% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_function.R
\name{find_corr}
\alias{find_corr}
\title{A Correlation Function}
\usage{
find_corr(year)
}
\arguments{
\item{year}{}
}
\description{
This function allows you to find the correlation matrix of NBA data for a specific year.
}
\examples{
find_corr()
}
\keyword{correlation}
