% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_function.R
\name{find_oldest}
\alias{find_oldest}
\title{An Age Function}
\usage{
find_oldest(year)
}
\arguments{
\item{year}{}
}
\description{
This function allows you to find the oldest NBA player with the highest points for a specific year.
}
\examples{
find_oldest()
}
\keyword{age}
