% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_function.R
\name{find_points}
\alias{find_points}
\title{A Points Function}
\usage{
find_points(year)
}
\arguments{
\item{year}{}
}
\description{
This function allows you to find the NBA player with the highest points for a specific year.
}
\examples{
find_points()
}
\keyword{points}
